+++ Makefile
@@ -87,7 +87,7 @@ ACNCLI_DIR = tools/acncli
 TELEMETRY_CONF_DIR = telemetry
 CNS_DIR = cns/service
 CNMS_DIR = cnms/service
-NPM_DIR = npm/plugin
+NPM_DIR = npm/cmd
 OUTPUT_DIR = output
 BUILD_DIR = $(OUTPUT_DIR)/$(GOOS)_$(GOARCH)
 IMAGE_DIR  = $(OUTPUT_DIR)/images
