+++ npm/cmd/cli/convertiptable.go
@@ -1,9 +1,9 @@
-package cmd
+package cli
 
 import (
 	"fmt"
 
-	"github.com/Azure/azure-container-networking/npm/debug/dataplane"
+	"github.com/Azure/azure-container-networking/npm/pkg/debug/dataplane"
 	"github.com/spf13/cobra"
 )
