+++ npm/cmd/cli/parseiptable.go
@@ -1,9 +1,9 @@
-package cmd
+package cli
 
 import (
 	"fmt"
 
-	"github.com/Azure/azure-container-networking/npm/debug/dataplane/parse"
+	"github.com/Azure/azure-container-networking/npm/pkg/debug/dataplane/parse"
 	"github.com/spf13/cobra"
 )
