--- npm/cmd/const.go
@@ -1,8 +1,0 @@
-package cmd
-
-import "errors"
-
-var (
-	errSrcNotSpecified = errors.New("source not specified")
-	errDstNotSpecified = errors.New("destination not specified")
-)
