+++ npm/config/config.go
@@ -1,0 +1,7 @@
+package config
+
+type NPMConfig struct {
+	ResyncPeriodInMinutes int    `json:"ResyncPeriodInMinutes"`
+	ListeningPort         int    `json:"ListeningPort"`
+	ListeningAddress      string `json:"ListeningAddress"`
+}
