+++ npm/pkg/debug/dataplane/trafficanalyzer.go
@@ -7,7 +7,7 @@ import (
 	"strings"
 
 	"github.com/Azure/azure-container-networking/npm"
-	"github.com/Azure/azure-container-networking/npm/debug/pb"
+	"github.com/Azure/azure-container-networking/npm/pkg/debug/pb"
 	"github.com/Azure/azure-container-networking/npm/util"
 	"google.golang.org/protobuf/encoding/protojson"
 )
