+++ npm/pkg/debug/testfiles/npmcache.json
